fixed symbolsTable types

+++ p5/inc/symbolsTable.h
@@ -85,7 +85,7 @@ bool st_set_scalar_variable(
   String identifier,
   DataType dt,
   Scope scope,
-  int_fast32_t pos
+  int pos
 );
 
 /**
@@ -103,8 +103,8 @@ bool st_set_vector_variable(
   String identifier,
   DataType dt,
   Scope scope,
-  int_fast32_t pos,
-  uint_fast32_t size
+  int pos,
+  uint_fast8_t size
 );
 
 /**
@@ -119,7 +119,7 @@ bool st_set_scalar_parametre(
   SymbolsTable *st,
   String identifier,
   DataType dt,
-  int_fast32_t pos
+  int pos
 );
 
 /**
@@ -135,8 +135,8 @@ bool st_set_vector_parametre(
   SymbolsTable *st,
   String identifier,
   DataType dt,
-  int_fast32_t pos,
-  uint_fast32_t size
+  int pos,
+  uint_fast8_t size
 );
 
 /**
@@ -177,8 +177,8 @@ bool st_insertBlindCurrentScope(
   DataType dataType,
   IdentifierCategory identifierCategory,
   Scope scope,
-  int_fast32_t pos,
-  uint_fast32_t size,
+  int pos,
+  uint_fast8_t size,
   int_fast32_t params,
   int_fast32_t localvars
 );
